requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.1.0
pkg-resources==0.0.0
python-dotenv==0.10.3
requests==2.22.0
urllib3==1.25.3
virtualenv==16.6.1
Werkzeug==0.15.4
